Fix error in CONVERT using character string into DATE or DATETIME (#621)
* Fix error in CONVERT using character string into DATE or DATETIME

Previously, CONVERT function fails when character string of format mm/dd/yy convert into DATE or DATETIME. The error is :operator is not unique: smallint > numeric. The error raised in ``babelfish_get_full_year`` function when trying to compare SMALLINT with NUMERIC. In this commit,  SMALLINT is converted into NUMERIC before doing comparison.

Task:BABEL-3314

+++ contrib/babelfishpg_tsql/sql/sys_function_helpers.sql
@@ -2185,7 +2185,7 @@ BEGIN
 
             v_full_year = v_base_century + v_short_year;
             v_full_year = CASE
-                             WHEN (v_short_year > p_year_cutoff) THEN v_full_year - 100
+                             WHEN (v_short_year::NUMERIC > p_year_cutoff) THEN v_full_year - 100
                              ELSE v_full_year
                           END;
         ELSE v_full_year := v_short_year;
